Add SetCurrentStack in up.go after running requireStack
This change, when run in the current Pulumi directory, will set the
backend state to remember the current stack.

TODO: set up situational unit tests that verify this is the behavior we
want.

Unclear why a stack name is not set on consecutive calls to `up` (see
q.Q debug output after running `pulumi up` again on an established
stack).