[resource] Add a generic NewProperty function
These changes add a generic package-level function to resource,
NewProperty, that creates a new PropertyValue from a plain value.

The primary value of these changes for the time being is the
introduction of the PropertyValueType interface, which documents the
possible types that may be contained in a PropertyValue.

In future changes I would like to unexport PropertyValue.V and replace
it with an exported method, `func (PropertyValue) Value() any`, that
prevents the consumer from sticking unexpected types in a
`PropertyValue`.

Someday maybe Go will have sum types and we can all be happy.