[sdk/python] Maintain old behavior for empty k8s invoke results
The hand-rolled invoke code in the Kubernetes Python SDK is not expecting empty dicts for empty results. Instead, it has code that checks for `None`.

In general, we want to keep the behavior of returning empty dicts rather than `None` for empty results, as this aligns with the behavior of other SDKs (Node.js and Go) and is actually the original behavior of the Python SDK (it regressed unintentionally in the v2 timeframe).

We're updating the Kubernetes Python SDK to be resilient to either getting an empty dict or `None` for invokes with empty results.

Additionally, to maintain compat with older versions of the Kubernetes Python SDK, we'll special case those invoke tokens. When an invoke's result is empty and the token is one of the Kubernete's function tokens, we'll return `None` rather than the empty dict.