Adds `pulumi package new`
This command adds a new plugin type, Boilerplate, which takes a name and config
(of unspecified format) and is expected to generate a new Pulumi package, i.e.:
a provider.

A working implementation of the Boilerplate provider is implemented in
pulumi/ci-mgmt#457. Using this command with the following configuration, a user
can generate a working Pulumi provider binary in a single command. As the
boilerplate is implemented as a plugin, it can be updated independently of the
Pulumi CLI, versioned, and programmatically shim bridged providers.

```yaml
module: github.com/pulumi/pulumi-porkbun
providerTitle: "Porkbun"
publisher: "Pulumi"
homepage: "https://www.pulumi.com/"
logoURL: ""
major-version: 0

upstreamOrg: cullenmcdermott
upstreamRepo: terraform-provider-porkbun

sdks:
  nodejs: "@pulumi/porkbun"
  python: pulumi_porkbun
  dotnet: Pulumi.Porkbun
  # Include the major version here, i.e.: sdk/v4 or just sdk for v0 and v1
  go: github.com/pulumi/pulumi-porkbun/sdk
```