Allow `import` to create empty component resources (#14467)
<!--- 
Thanks so much for your contribution! If this is your first time
contributing, please ensure that you have read the
[CONTRIBUTING](https://github.com/pulumi/pulumi/blob/master/CONTRIBUTING.md)
documentation.
-->

# Description

<!--- Please include a summary of the change and which issue is fixed.
Please also include relevant motivation and context. -->

Fixes https://github.com/pulumi/pulumi/issues/14444.

This adds two new flags to the import file to signal that a resource is
a component, and also if that is a remote (MLC) component. The import
system will create empty resources for those import descriptions, but
importantly they can then be used as parents for other resources in the
import deployment.

We need to know if it's a remote component to know if a provider should
be looked up for it. We could probably get away with _just_ knowing if
it's a component at import time and not doing any validation that the
type token is well formed and the provider exists for MLCs. That makes
import a little simpler for users (no need to set if it's remote or not)
but pushes error cases downstream to when they actually try to `up`
their program. It also means we can't lookup and fill in the default
providers for these resources, resulting in a provider diff at `up`
time.

## Checklist

- [x] I have run `make tidy` to update any new dependencies
- [x] I have run `make lint` to verify my code passes the lint check
  - [ ] I have formatted my code using `gofumpt`

<!--- Please provide details if the checkbox below is to be left
unchecked. -->
- [x] I have added tests that prove my fix is effective or that my
feature works
<!--- 
User-facing changes require a CHANGELOG entry.
-->
- [x] I have run `make changelog` and committed the
`changelog/pending/<file>` documenting my change
<!--
If the change(s) in this PR is a modification of an existing call to the
Pulumi Cloud,
then the service should honor older versions of the CLI where this
change would not exist.
You must then bump the API version in
/pkg/backend/httpstate/client/api.go, as well as add
it to the service.
-->
- [ ] Yes, there are changes in this PR that warrants bumping the Pulumi
Cloud API version
<!-- @Pulumi employees: If yes, you must submit corresponding changes in
the service repo. -->