add output for running policy packs (#14493)
# Description

:construction: 

This is a work in progress attempting to add some output for when policy
packs are being loaded/compiled (which I realized was the actual reason
for the slowdown, not that they were being run already). There's still
some pretty obvious code quality issues as I've been throwing this
together, and the URN's are not constructed properly yet either, so some
of the output is less helpful than it could be. I'm mainly opening this
to get some early feedback on whether this is the appropriate place to
add the output, or if there are better alternatives.

Also including a gif here of what this looks like. In particular the
Name here needs to be improved, and the "Plan" might want to say
"loading/compiling" or something similar instead?


![policy-pack](https://github.com/pulumi/pulumi/assets/191004/85cca146-d740-4e4c-96dc-cc26dca6bb90)

Fixes #14453

## Checklist

- [ ] I have run `make tidy` to update any new dependencies
- [ ] I have run `make lint` to verify my code passes the lint check
  - [ ] I have formatted my code using `gofumpt`

<!--- Please provide details if the checkbox below is to be left
unchecked. -->
- [ ] I have added tests that prove my fix is effective or that my
feature works
<!--- 
User-facing changes require a CHANGELOG entry.
-->
- [ ] I have run `make changelog` and committed the
`changelog/pending/<file>` documenting my change
<!--
If the change(s) in this PR is a modification of an existing call to the
Pulumi Cloud,
then the service should honor older versions of the CLI where this
change would not exist.
You must then bump the API version in
/pkg/backend/httpstate/client/api.go, as well as add
it to the service.
-->
- [ ] Yes, there are changes in this PR that warrants bumping the Pulumi
Cloud API version
<!-- @Pulumi employees: If yes, you must submit corresponding changes in
the service repo. -->