Add a strict opinionated promise library (#14552)
<!--- 
Thanks so much for your contribution! If this is your first time
contributing, please ensure that you have read the
[CONTRIBUTING](https://github.com/pulumi/pulumi/blob/master/CONTRIBUTING.md)
documentation.
-->

# Description

<!--- Please include a summary of the change and which issue is fixed.
Please also include relevant motivation and context. -->

Prompted by having another case in
https://github.com/pulumi/pulumi/pull/14548 that looked like "this is
just a promise".

We already had a mini-promise library for provider config in
sdk/go/common/resource/plugin/provider_plugin.go. Plus a couple of
places where we did a poor mans Go promise by having a WaitGroup plus a
variable to set (see the two changes in /pkg).

This adds a little promise library to safely cover all three of these
cases plus the case I'll be adding in #14548. It is _minimal_ adding
just the features of promises needed for these initial cases. We can add
to it as needed.

I suspect we've got a number of places in test code that could probably
use this as well, but I haven't gone through that. I also suspect that
having this type available will result in more places in the future
being simpler because "its just a promise" is a fairly common scenario
in async systems. In fact Output's internally _are just a promise_ so we
could probably rewrite their internals using this, add a `Then()` like
method for apply and all the async complexity gets handled in the
promise layer while the output layer just cares about
unknowns/secrets/etc.

## Checklist

- [x] I have run `make tidy` to update any new dependencies
- [x] I have run `make lint` to verify my code passes the lint check
  - [ ] I have formatted my code using `gofumpt`

<!--- Please provide details if the checkbox below is to be left
unchecked. -->
- [x] I have added tests that prove my fix is effective or that my
feature works
<!--- 
User-facing changes require a CHANGELOG entry.
-->
- [ ] I have run `make changelog` and committed the
`changelog/pending/<file>` documenting my change - This is in
sdk/go/common and I'm considering it an non-public api for now.
<!--
If the change(s) in this PR is a modification of an existing call to the
Pulumi Cloud,
then the service should honor older versions of the CLI where this
change would not exist.
You must then bump the API version in
/pkg/backend/httpstate/client/api.go, as well as add
it to the service.
-->
- [ ] Yes, there are changes in this PR that warrants bumping the Pulumi
Cloud API version
<!-- @Pulumi employees: If yes, you must submit corresponding changes in
the service repo. -->