[go/sdk-gen] Fix generics-only option missing ToOutput(...) methods (#14584)
# Description

PR #14492 removed `ToOutput(...)` methods for output types when
generating non-generic SDK variants. The fix there was added to only
include these when `side-by-side` is enabled, which implicitly included
`generics-only` except for a _single_ place where that wasn't the case 😓
(see comment below in `gen.go`)

This PR fixes issue and includes tests for go sdkgen where `generics`
setting is set to `generics-only`:
 - `output-funcs-go-generics-only`
 - `plain-and-default-go-generics-only`
 - `secrets-go-generics-only`
 - `simple-enum-schema-go-generics-only`
 - `simple-plain-schema-go-generics-only`

## Checklist

- [ ] I have run `make tidy` to update any new dependencies
- [x] I have run `make lint` to verify my code passes the lint check
  - [x] I have formatted my code using `gofumpt`

<!--- Please provide details if the checkbox below is to be left
unchecked. -->
- [x] I have added tests that prove my fix is effective or that my
feature works
<!--- 
User-facing changes require a CHANGELOG entry.
-->
- [x] I have run `make changelog` and committed the
`changelog/pending/<file>` documenting my change
<!--
If the change(s) in this PR is a modification of an existing call to the
Pulumi Cloud,
then the service should honor older versions of the CLI where this
change would not exist.
You must then bump the API version in
/pkg/backend/httpstate/client/api.go, as well as add
it to the service.
-->
- [ ] Yes, there are changes in this PR that warrants bumping the Pulumi
Cloud API version
<!-- @Pulumi employees: If yes, you must submit corresponding changes in
the service repo. -->