Use promise rather than `sync.Value` to record step errors.
Fixes https://github.com/pulumi/pulumi/issues/14611.

`sync.Value` panics if a second value is written to it which isn't
_exactly_ the same type. We were using this to atomicly record any
errors that happened, although where only interested in saving the first
error.

This was fine if no or one error happened, or if multiple errors
happened that all were the same error type. But if two errors if
different error types happened `sync.Value` would panic.

Switched to use a promise completion source instead. If any error
happens we use it to reject the completion source, we also log if this
isn't the first error we've seen (a capabilitiy that `sync.Value` didn't
give us). At the end of the step executor being used when `Errored()` is
called we try to get the result of the completion source.