Skip TestLanguageNewSmoke when not on .NET 8+
The templates for `pulumi new` were updated to target .NET 8, the new LTS version of .NET. But our test matrix runs on both .NET 6 and .NET 8, which is causing `TestLanguageNewSmoke` to fail when running on .NET 6 because it doesn't support targeting .NET 8. To mitigate, we'll skip running this test when the version of `dotnet` is not 8 or greater.