[esc] Add a command to create an ESC env from stack config (#14634)
This change adds a command, `pulumi config env init`, that creates a new
environment for a stack based on the stack's configuration values.

From the usage:

* `pulumi config env init`

Creates an environment for a specific stack based on the stack's
configuration values, then replaces the stack's configuration values
with a reference to that environment. The environment will be created in
the same organization as the stack.