Fix test-language to not use the default loader for binding (#14656)
Because BindSpec was being passed nil for the loader parameter it was
creating a new default plugin host, instead of using the mocked one we
create as part of the test-language code.

No real observable effect as yet, but would of been an issue once we
start having test schemas that refer to each other.