Fix some lint issues in plugins_install_test.go (#14644)
This isn't currently getting picked up by the linter because it's behind
a build tag and the linter doesn't run with those set.

It _should_ because it turns out loads of test files aren't being linted
right now, but there's quite a few fixes that need to happen so linting
doesn't then fail, so this is just one step towards that.