Add `--suppress-progresss` flag to CLI
Adds support for suppressing the periodic "..." printing that can disrupt normal output stream.  This output is still deemed necessary to include by default for CI systems that might otherwise cancel an update that goes for a long time without printing anything.  But we now have an option to turn this off.